split .h to .cc
