+++ caffe2/serialize/CMakeLists.txt
@@ -1,4 +1,7 @@
-file(GLOB tmp *_test.cc)
-set(Caffe2_CPU_TEST_SRCS ${Caffe2_CPU_TEST_SRCS} ${tmp})
+set(Caffe2_CPU_SRCS ${Caffe2_CPU_SRCS}
+  serialize/intermediate_model.cc)
+set(Caffe2_CPU_TEST_SRCS ${Caffe2_CPU_TEST_SRCS}
+  serialize/intermediate_model_test.cc)
 
+set(Caffe2_CPU_SRCS ${Caffe2_CPU_SRCS} PARENT_SCOPE)
 set(Caffe2_CPU_TEST_SRCS ${Caffe2_CPU_TEST_SRCS} PARENT_SCOPE)
